backend/requirements.txt
qwen_vl_utils
youtube-transcript-api>=1.2.4
google-generativeai
python-dotenv
uvicorn
fastapi
yt-dlp
opencv-python
transformers